backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
langgraph
langchain
langchain_groq
python-dotenv
pydantic
requests
